### 2025-02-20 #7818 @editorialbot
[PRE REVIEW]: pengWann: Descriptors of chemical bonding from Wannier functions
**Submitting author:** @PatrickJTaylor (<a href="http://orcid.org/0009-0003-6511-6442">Patrick J. Taylor</a>)
**Repository:** https://github.com/PatrickJTaylor/pengWann
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.1
