### 2025-02-20 #7818 @PatrickJTaylor
Hi!

Having looked at the list of registered JOSS reviewers, here are a few suggestions that I think may be suitable:

- JosePizarro3
- berquist
- ChrisWoodgate
- r2stanton
- HugoStrand
