### 2025-02-20 ec8d99a
Site is now a dataclass rather than a NamedTuple
