### 2025-02-20 ab0f74a
Simplify logic of identify_interatomic_interactions
