### 2025-02-20 f9fcce9
Refer to zero_vector rather than bl_0 for readability
