### 2025-02-20 dc3d8af
Remove unused class attribute from DescriptorCalculator
