### 2025-02-20 8e687a6
Add __getitem__ and __str__ to Geometry
