### 2025-02-20 7f6f132
Only call distance matrix after checking for Wannier centres in assignments
