+++ tests/test_geometry.py
@@ -13,18 +13,29 @@
 # You should have received a copy of the GNU General Public License along with pengWann.
 # If not, see <https://www.gnu.org/licenses/>.
 
-import json
 import pytest
 import numpy as np
 from pengwann.geometry import (
     AtomicInteraction,
-    assign_wannier_centres,
-    build_geometry,
+    Geometry,
     identify_interatomic_interactions,
     identify_onsite_interactions,
+    Site,
 )
-from pymatgen.analysis.structure_matcher import StructureMatcher
-from pymatgen.core import Structure
+
+
+def build_geometry(symbols: list[str]) -> Geometry:
+    cell = np.diag([5, 5, 5])
+
+    coords = np.array(
+        [[0.1, 0.1, 0.1], [0.6, 0.6, 0.6], [0.25, 0.25, 0.25], [0.75, 0.75, 0.75]]
+    )
+    sites = tuple(
+        Site(symbol, idx, coords)
+        for idx, (symbol, coords) in enumerate(zip(symbols, coords))
+    )
+
+    return Geometry(sites, cell)
 
 
 def serialise_interactions(
@@ -48,78 +59,116 @@ def serialise_interactions(
 
 
 @pytest.fixture
-def geometry() -> Structure:
-    cell = np.diag([5, 5, 5])
-    species = ["X0+", "X0+", "C", "O"]
-    coords = [[0.1, 0.1, 0.1], [0.6, 0.6, 0.6], [0.25, 0.25, 0.25], [0.75, 0.75, 0.75]]
-    geometry = Structure(cell, species, coords)
+def geometry() -> Geometry:
+    symbols = ["X", "X", "C", "O"]
 
-    wannier_centres = ((2,), (3,), (0,), (1,))
-    geometry.add_site_property("wannier_centres", wannier_centres)
-
-    return geometry
+    return build_geometry(symbols)
 
 
-def test_build_geometry(shared_datadir) -> None:
-    test_geometry = build_geometry("wannier90", f"{shared_datadir}")
-    num_wann = len([site for site in test_geometry if site.species_string == "X0+"])
+@pytest.fixture
+def geometry_elemental() -> Geometry:
+    symbols = ["X", "X", "C", "C"]
 
-    with open(f"{shared_datadir}/geometry.json", "r") as stream:
-        serial = json.load(stream)
-
-    ref_geometry = Structure.from_dict(serial)
-
-    sm = StructureMatcher()
-
-    assert num_wann == 8
-    assert sm.fit(test_geometry, ref_geometry)
+    return build_geometry(symbols)
 
 
-def test_build_geometry_with_cell(shared_datadir) -> None:
+@pytest.fixture
+def geometry_no_x() -> Geometry:
+    symbols = ["C", "C", "C", "O"]
+
+    return build_geometry(symbols)
+
+
+def test_Geometry_from_xyz(
+    shared_datadir, data_regression, ndarrays_regression, tol
+) -> None:
+    geometry = Geometry.from_xyz("wannier90", f"{shared_datadir}")
+
+    symbols, indices, coords_list = [], [], []
+    for site in geometry:
+        symbols.append(site.symbol)
+        indices.append(site.index)
+        coords_list.append(site.coords)
+
+    coords = np.vstack(coords_list)
+
+    data_regression.check({"symbols": symbols, "indices": indices})
+    ndarrays_regression.check(
+        {"cell": geometry.cell, "coords": coords}, default_tolerance=tol
+    )
+
+
+def test_Geometry_from_xyz_with_cell(
+    shared_datadir, data_regression, ndarrays_regression, tol
+) -> None:
     cell = (
         (-1.7803725545451619, -1.7803725545451616, 0.0000000000000000),
         (-1.7803725545451616, 0.0000000000000000, -1.7803725545451616),
[…]
