+++ tests/test_geometry/test_Geometry_wannier_assignments.yml
@@ -1,4 +1,4 @@
-wannier_centres:
+wannier_assignments:
 - - 2
 - - 3
 - - 0
