+++ tests/test_geometry/test_Geometry_str.yml
@@ -1,0 +1,2 @@
+str: "Geometry\n========\nCell\n----\n[[5 0 0]\n [0 5 0]\n [0 0 5]]\n\nSites\n-----\n\
+  X0 [0.1 0.1 0.1]\nX1 [0.6 0.6 0.6]\nC2 [0.25 0.25 0.25]\nO3 [0.75 0.75 0.75]\n"
