### 2025-02-20 dca3fde
Add docstrings to new objects
