+++ src/pengwann/geometry.py
@@ -39,6 +39,29 @@ from pengwann.io import read_cell, read_xyz
 
 @dataclass(frozen=True)
 class Geometry:
+    """
+    Data structure representing a periodic structure of atoms and Wannier centres.
+
+    Attributes
+    ----------
+    sites : tuple of Site
+        The individual Site objects representing each atom/Wannier centre in the
+        structure.
+    cell : ndarray of float
+        The cell vectors associated with the structure.
+    wannier_assignments : tuple of tuple of int
+    distance_and_image_matrices : tuple of ndarray of float
+
+    See Also
+    --------
+    Site
+
+    Notes
+    -----
+    This class should not normally be initialised using the base constructor. See
+    instead the :py:meth:`~pengwann.geometry.Geometry.from_xyz` classmethod.
+    """
+
     sites: tuple[Site, ...]
     cell: NDArray[np.float64]
 
@@ -70,6 +93,23 @@ class Geometry:
 
     @cached_property
     def wannier_assignments(self) -> tuple[tuple[int, ...], ...]:
+        """
+        Assign Wannier centres to atoms based on a closest distance criterion.
+
+        The indices returned by this property refer to the order of sites in the
+        calling Geometry object. Each Wannier centre is associated with a single index
+        identifying its closest atom whilst each atom is associated with a sequence of
+        indices identifying all of its assigned Wannier centres.
+
+        Returns
+        -------
+        wannier_assignments : tuple of tuple of int
+            The indices assigning Wannier centres to atoms.
+
+        Notes
+        -----
+        This property is cached once it has been evaluated once.
+        """
         wannier_indices, atom_indices = [], []
         for site in self.sites:
             if site.symbol == "X":
@@ -99,6 +139,25 @@ class Geometry:
     def distance_and_image_matrices(
         self,
     ) -> tuple[NDArray[np.float64], NDArray[np.int_]]:
+        """
+        Compute the distance and image matrices.
+
+        The image matrix mirrors the shape of the distance matrix, with each element
+        referring to a Bravais lattice vector that specifies the periodic image of a
+        given site that is closest to another site.
+
+        Returns
+        -------
+        distance_matrix : ndarray of float
+            The distance matrix.
+
+        image_matrix : ndarray of np.int_
+            The image matrix.
+
+        Notes
+        -----
+        This property is cached once it has been evaluated once.
+        """
         num_sites = len(self)
         num_dim = len(self.cell)
         distance_matrix = np.zeros((num_sites, num_sites))
@@ -132,6 +191,38 @@ class Geometry:
     def from_xyz(
         cls, seedname: str, path: str = ".", cell: ArrayLike | None = None
     ) -> Geometry:
+        """
+        Initialise a Geometry object from a seedname_centres.xyz file.
+
+        Parameters
+        ----------
+        seedname : str
+            The seedname (prefix for all input files) chosen in the prior Wannier90
+            calculation.
+        path : str
+            Filepath to the directory containing seedname_centres.xyz and (optionally)
+            seedname.win. Defaults to '.' i.e. the current working directory.
+        cell : array_like or None, optional
+            The cell vectors associated with the structure. Defaults to None, in which
+            case the cell vectors will automatically be extracted from a seedname.win
+            file in the same directory as the xyz file.
+
+        Returns
+        -------
+        geometry : Geometry
+            The initialised Geometry object.
+
+        Notes
+        -----
+        If you intend to use the instance returned by this method as an input to the
+        :py:func:`~pengwann.geometry.identify_interatomic_interactions` function, please
+        ensure that the seedname_centres.xyz file was generated by Wannier90 with
[…]
