### 2025-02-20 b333e28
Print Wannier assignments from __str__ method of Geometry
