### 2025-02-20 4b572b6
Update test for __str__ method of Geometry
