### 2025-02-20 98a0ab9
Update old docstrings in geometry module
