### 2025-02-20 46a65c1
Fix minor typos in docstrings of geometry module
