### 2025-02-20 a8f1214
Hacky workaround for numpydoc not linking cached_property
