### 2025-02-20 47f30ce
Update rst for new geometry module docs
