+++ docs/utils.rst
@@ -11,6 +11,5 @@ Functions
    :toctree: generated
 
    allocate_shared_memory
-   get_atom_indices
    get_spilling_factor
    integrate_descriptor
