### 2025-02-20 ba5f14f
Add as_structure method and update docs of Geometry class
