+++ src/pengwann/geometry.py
@@ -35,12 +35,22 @@ from pengwann.interactions import (
     WannierInteraction,
 )
 from pengwann.io import read_cell, read_xyz
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from pymatgen.core import Structure
 
 
 @dataclass(frozen=True)
 class Geometry:
     """
     Data structure representing a periodic system of atoms and Wannier centres.
+
+    The primary function of this class is to store the information required to identify
+    interatomic and on-site interactions in terms of individual Wannier functions. In
+    addition, for ease of integration into common materials modelling workflows,
+    Geometry objects can be easily converted into Pymatgen Structure objects with the
+    :py:meth:`~pengwann.geometry.Geometry.as_structure` method.
 
     Attributes
     ----------
@@ -60,6 +70,17 @@ class Geometry:
     -----
     This class should not normally be initialised using the base constructor. See
     instead the :py:meth:`~pengwann.geometry.Geometry.from_xyz` classmethod.
+
+    Examples
+    --------
+    To access individual Site objects:
+
+    >>> site_0 = geometry[0]
+
+    To iterate over all sites:
+
+    >>> for site in geometry:
+    ...     # Do something for each Site object.
     """
 
     sites: tuple[Site, ...]
@@ -198,6 +219,48 @@ class Geometry:
 
         return distance_matrix, image_matrix
 
+    def as_structure(self) -> Structure:
+        """
+        Build a Pymatgen Structure object from the calling Geometry object.
+
+        Returns
+        -------
+        structure : Structure
+            The Pymatgen Structure object.
+
+        Notes
+        -----
+        The :py:meth:`~pengwann.geometry.Geometry.wannier_assignments` of the Geometry
+        object are preserved in the output Pymatgen Structure as a
+        :code:`site_property`.
+
+        Examples
+        --------
+        To access the Wannier assignments:
+
+        >>> assignments = structure.site_properties["wannier_assignments"]
+        """
+        try:
+            from pymatgen.core import Structure
+
+        except ImportError as base_error:
+            raise ImportError(
+                """The as_structure method requires the Pymatgen package, which cannot
+                be found."""
+            ) from base_error
+
+        symbols, coords = [], []
+        for site in self.sites:
+            symbols.append(site.symbol)
+            coords.append(site.coords)
+
+        return Structure(
+            lattice=self.cell,
+            species=symbols,
+            coords=coords,
+            site_properties={"wannier_assignments": self.wannier_assignments},
+        )
+
     @classmethod
     def from_xyz(
         cls, seedname: str, path: str = ".", cell: ArrayLike | None = None
@@ -225,9 +288,9 @@ class Geometry:
 
         Notes
         -----
-        If you intend to use the instance returned by this method as an input to the
-        :py:func:`~pengwann.geometry.identify_interatomic_interactions` function, please
-        ensure that the seedname_centres.xyz file was generated by Wannier90 with
+        If you intend to use the Geometry object returned by this method as an input to
+        the :py:func:`~pengwann.geometry.identify_interatomic_interactions` function,
+        please ensure that the seedname_centres.xyz file was generated by Wannier90 with
         translate_home_cell = false. If the seedname_centres.xyz file was generated with
         translate_home_cell = true, then the interactions identified by the
         :py:func:`~pengwann.geometry.identify_interatomic_interactions` function may not
