### 2025-02-20 24ba439
Add test for as_structure method
