### 2025-02-20 dd23664
Pymatgen core => dev dependency
