### 2025-02-21 02795b3
Further edits to docstrings
