### 2025-02-21 d7460c0
Add JOSS badge to README and slightly rephrase summary
