### 2025-02-21 25b775c
Revert "Test example outputs in README"
This reverts commit 76e6a34f19d223e23763cf07c37fde0a8405c195.
