### 2025-02-21 e366089
Make README more verbose/clear
