### 2025-02-21 5d9bf0a
Remove github emojis from README (as they do not display on PyPI)
