### 2025-02-21 45ad315
Use footnote rather than formal ref in README
