### 2025-02-21 380a962
Update example outputs svg with transparency and add to README
