### 2025-02-21 4fb2132
Rename example outputs svg for compatibility with README
