### 2025-02-21 5de55af
Fix typo in README
