+++ README.md
@@ -23,7 +23,7 @@ A lightweight Python package for computing descriptors of chemical bonding and l
 <p align="center">
   <small>
   A handful of example outputs from <code>pengwann</code> as applied to rutile.
-  The colour-coded numbers next to the crystal structure are Lwdin-style charges computed for Ti (blue) and O (red).
+  The colour-coded numbers next to the crystal structure are Löwdin-style charges computed for Ti (blue) and O (red).
   </small>
 </p>
 
