### 2025-02-21 fe4b241
Tweak features summary in README
