### 2025-02-27 5ddcdb5
Add rust extension for building distance and image matrices
