### 2025-02-28 110b8d1
Update test suite to account for rust functions
