### 2025-02-28 b8d2677
Fix enum_image_vectors rust function and sort use statements
