### 2025-02-28 af0718b
Attempt to adapt CI to new rust code
