+++ .github/workflows/ci.yml
@@ -19,13 +19,26 @@ jobs:
           python-version: ${{ matrix.python-version }}
           enable-cache: true
 
+      - name: Install Rust toolchain
+        uses: dtolnay/rust-toolchain@stable
+        with:
+          components: clippy, rustfmt
+
       - name: Sync environment
         run: uv sync
 
-      - name: Lint
+      - name: Lint Rust
+        working-directory: rust
+        run: cargo clippy
+
+      - name: Enforce Rust formatting
+        working-directory: rust
+        run: cargo fmt --check
+
+      - name: Lint Python
         run: uv run ruff check
 
-      - name: Enforce black formatting
+      - name: Enforce Python formatting
         run: uv run ruff format --check
 
       - name: Run static type checking
