### 2025-02-28 cba34b5
Setup Rust before Python in CI
