### 2025-02-28 ac085d4
Limit version of scikit_build_core as temporary workaround for spglib build
