+++ pyproject.toml
@@ -82,6 +82,7 @@ dev = [
     "pytest-random-order>=1.1.1",
     "pytest-regressions>=2.7.0",
     "ruff>=0.9.5",
+    "scikit-build-core<=0.10.7",
     "scipy-stubs>=1.15.1.0",
     "types-tqdm>=4.67.0.20241221",
 ]
