### 2025-02-28 c03636d
Attempt to change order of dependency installation re: spglib build
