### 2025-02-28 ee81db0
Remove scikit_build_core dev dependency (workaround does not work)
