+++ pyproject.toml
@@ -74,7 +74,6 @@ testpaths = ['tests']
 
 [dependency-groups]
 dev = [
-    "scikit-build-core<0.11",
     "pymatgen>=2025.1.24",
     "pyright>=1.1.393",
     "pytest>=8.3.4",
