### 2025-02-28 d97f49a
Remove now unused imports from geometry module
