### 2025-02-28 d0b3f0d
Split linting/formatting/typing away from testing in CI
