+++ .github/workflows/ci.yml
@@ -3,12 +3,9 @@ name: Continuous integration
 on: [push, pull_request]
 
 jobs:
-  run:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ['3.10', '3.11', '3.12']
+  check:
+    name: Lint, format and type check
+    runs-on: ubuntu-latest
 
     steps:
       - name: Checkout repository
@@ -17,7 +14,6 @@ jobs:
       - name: Install uv and Python
         uses: astral-sh/setup-uv@v5
         with:
-          python-version: ${{ matrix.python-version }}
           enable-cache: true
 
       - name: Install Rust toolchain
@@ -42,14 +38,41 @@ jobs:
       - name: Enforce Python formatting
         run: uv run ruff format --check
 
-      - name: Run static type checking
+      - name: Run static type checker
         run: uv run pyright
+
+  test:
+    name: Run test suite and publish to Code Climate
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ['3.10', '3.11', '3.12']
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Install uv and Python
+        uses: astral-sh/setup-uv@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+          enable-cache: true
+
+      - name: Install Rust toolchain
+        uses: dtolnay/rust-toolchain@stable
+        with:
+          components: clippy, rustfmt
+
+      - name: Sync environment
+        run: uv sync
 
       - name: Run test suite
         run: uv run pytest
 
       - name: Publish test coverage report
         if: github.event_name == 'push'
+        if: ${{ (github.event_name == 'push') && (runner.os == 'Linux') && (matrix.python-version == '3.12') }}
         uses: paambaati/codeclimate-action@v9.0.0
         env:
           CC_TEST_REPORTER_ID: ${{ secrets.CODECLIMATE_REPO_TOKEN }}
