### 2025-02-28 b5ba3f3
Remove duplicate conditional from CI
