+++ .github/workflows/release.yml
@@ -1,31 +1,141 @@
 name: Release
 
-on:
-  push:
-    tags:
-      - v*
+#  on:
+#    push:
+#      tags:
+#        - v*
+#  
+#  jobs:
+#    run:
+#      name: Build and publish to PyPI
+#      runs-on: ubuntu-latest
+#      environment:
+#        name: pypi
+#      permissions:
+#        id-token: write
+#  
+#      steps:
+#        - name: Checkout repository
+#          uses: actions/checkout@v4
+#  
+#        - name: Install uv and Python
+#          uses: astral-sh/setup-uv@v5
+#          with:
+#            python-version: "3.12"
+#            enable-cache: true
+#  
+#        - name: Build wheel and sdist
+#          run: uv build
+#  
+#        - name: Publish to PyPI
+#          run: uv publish
+
+on: push
 
 jobs:
-  run:
-    name: Build and publish to PyPI
+  linux:
     runs-on: ubuntu-latest
-    environment:
-      name: pypi
-    permissions:
-      id-token: write
+    strategy:
+      matrix:
+        python-version: ["3.10", "3.11", "3.12"]
+        target: [x86_64, x86, aarch64, armv7, s390x, ppc64le]
 
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
 
-      - name: Install uv and Python
-        uses: astral-sh/setup-uv@v5
+      - name: Install Python
+        uses: actions/setup-python@v5
         with:
-          python-version: "3.12"
-          enable-cache: true
+          python-version: ${{ matrix.python-version }}
 
-      - name: Build wheel and sdist
-        run: uv build
+      - name: Build wheels
+        uses: PyO3/maturin-action@v1
+        with:
+          target: ${{ matrix.target }}
+          args: --release --out dist --find-interpreter --manifest-path rust/Cargo.toml
+          sccache: 'true'
+          manylinux: auto
 
-      - name: Publish to PyPI
-        run: uv publish
+      - name: Upload wheels
+        uses: actions/upload-artifact@v3
+        with:
+          name: wheels
+          path: dist
+
+  macos:
+    runs-on: macos-latest
+    strategy:
+      matrix:
+        python-version: ["3.10", "3.11", "3.12"]
+        target: [x86_64, aarch64]
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Install Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Build wheels
+        uses: PyO3/maturin-action@v1
+        with:
+          target: ${{ matrix.target }}
+          args: --release --out dist --find-interpreter --manifest-path rust/Cargo.toml
+          sccache: 'true'
+
+      - name: Upload wheels
+        uses: actions/upload-artifact@v3
+        with:
+          name: wheels
+          path: dist
+
+  windows:
+    runs-on: windows-latest
+    strategy:
+      matrix:
+        python-version: ["3.10", "3.11", "3.12"]
+        target: [x64, x86]
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Install Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Build wheels
+        uses: PyO3/maturin-action@v1
+        with:
+          target: ${{ matrix.target }}
+          args: --release --out dist --find-interpreter --manifest-path rust/Cargo.toml
+          sccache: 'true'
+
+      - name: Upload wheels
+        uses: actions/upload-artifact@v3
+        with:
+          name: wheels
+          path: dist
+
+  sdist:
[…]
