### 2025-02-28 4a02725
Test maturin-generated release workflow
