+++ .github/workflows/release.yml
@@ -4,7 +4,7 @@ name: Release
 #    push:
 #      tags:
 #        - v*
-#  
+#
 #  jobs:
 #    run:
 #      name: Build and publish to PyPI
@@ -13,20 +13,20 @@ name: Release
 #        name: pypi
 #      permissions:
 #        id-token: write
-#  
+#
 #      steps:
 #        - name: Checkout repository
 #          uses: actions/checkout@v4
-#  
+#
 #        - name: Install uv and Python
 #          uses: astral-sh/setup-uv@v5
 #          with:
 #            python-version: "3.12"
 #            enable-cache: true
-#  
+#
 #        - name: Build wheel and sdist
 #          run: uv build
-#  
+#
 #        - name: Publish to PyPI
 #          run: uv publish
 
@@ -34,108 +34,139 @@ on: push
 
 jobs:
   linux:
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.platform.runner }}
     strategy:
       matrix:
         python-version: ["3.10", "3.11", "3.12"]
-        target: [x86_64, x86, aarch64, armv7, s390x, ppc64le]
-
+        platform:
+          - runner: ubuntu-22.04
+            target: x86_64
+          - runner: ubuntu-22.04
+            target: x86
+          - runner: ubuntu-22.04
+            target: aarch64
+          - runner: ubuntu-22.04
+            target: armv7
+          - runner: ubuntu-22.04
+            target: s390x
+          - runner: ubuntu-22.04
+            target: ppc64le
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Install Python
-        uses: actions/setup-python@v5
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-
       - name: Build wheels
         uses: PyO3/maturin-action@v1
         with:
-          target: ${{ matrix.target }}
-          args: --release --out dist --find-interpreter --manifest-path rust/Cargo.toml
-          sccache: 'true'
+          target: ${{ matrix.platform.target }}
+          args: --release --out dist --find-interpreter
+          sccache: true
           manylinux: auto
-
       - name: Upload wheels
         uses: actions/upload-artifact@v4
         with:
-          name: wheels
+          name: wheels-linux-${{ matrix.platform.target }}
           path: dist
 
-  macos:
-    runs-on: macos-latest
+  musllinux:
+    runs-on: ${{ matrix.platform.runner }}
     strategy:
       matrix:
         python-version: ["3.10", "3.11", "3.12"]
-        target: [x86_64, aarch64]
-
+        platform:
+          - runner: ubuntu-22.04
+            target: x86_64
+          - runner: ubuntu-22.04
+            target: x86
+          - runner: ubuntu-22.04
+            target: aarch64
+          - runner: ubuntu-22.04
+            target: armv7
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Install Python
-        uses: actions/setup-python@v5
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
-
       - name: Build wheels
         uses: PyO3/maturin-action@v1
         with:
-          target: ${{ matrix.target }}
-          args: --release --out dist --find-interpreter --manifest-path rust/Cargo.toml
-          sccache: 'true'
-
+          target: ${{ matrix.platform.target }}
+          args: --release --out dist --find-interpreter
+          sccache: true
+          manylinux: musllinux_1_2
       - name: Upload wheels
         uses: actions/upload-artifact@v4
         with:
-          name: wheels
+          name: wheels-musllinux-${{ matrix.platform.target }}
[…]
