### 2025-02-28 20a5ca8
Add maturin as a dev dependency for convenience
