### 2025-02-28 b3f101c
Use default maturin CI config
