+++ .github/workflows/release.yml
@@ -1,35 +1,5 @@
 name: Release
 
-#  on:
-#    push:
-#      tags:
-#        - v*
-#
-#  jobs:
-#    run:
-#      name: Build and publish to PyPI
-#      runs-on: ubuntu-latest
-#      environment:
-#        name: pypi
-#      permissions:
-#        id-token: write
-#
-#      steps:
-#        - name: Checkout repository
-#          uses: actions/checkout@v4
-#
-#        - name: Install uv and Python
-#          uses: astral-sh/setup-uv@v5
-#          with:
-#            python-version: "3.12"
-#            enable-cache: true
-#
-#        - name: Build wheel and sdist
-#          run: uv build
-#
-#        - name: Publish to PyPI
-#          run: uv publish
-
 on: push
 
 jobs:
@@ -37,7 +7,6 @@ jobs:
     runs-on: ${{ matrix.platform.runner }}
     strategy:
       matrix:
-        python-version: ["3.10", "3.11", "3.12"]
         platform:
           - runner: ubuntu-22.04
             target: x86_64
@@ -55,7 +24,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.python-version }}
+          python-version: 3.x
       - name: Build wheels
         uses: PyO3/maturin-action@v1
         with:
@@ -73,7 +42,6 @@ jobs:
     runs-on: ${{ matrix.platform.runner }}
     strategy:
       matrix:
-        python-version: ["3.10", "3.11", "3.12"]
         platform:
           - runner: ubuntu-22.04
             target: x86_64
@@ -87,7 +55,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.python-version }}
+          python-version: 3.x
       - name: Build wheels
         uses: PyO3/maturin-action@v1
         with:
@@ -105,7 +73,6 @@ jobs:
     runs-on: ${{ matrix.platform.runner }}
     strategy:
       matrix:
-        python-version: ["3.10", "3.11", "3.12"]
         platform:
           - runner: windows-latest
             target: x64
@@ -115,7 +82,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.python-version }}
+          python-version: 3.x
           architecture: ${{ matrix.platform.target }}
       - name: Build wheels
         uses: PyO3/maturin-action@v1
@@ -133,7 +100,6 @@ jobs:
     runs-on: ${{ matrix.platform.runner }}
     strategy:
       matrix:
-        python-version: ["3.10", "3.11", "3.12"]
         platform:
           - runner: macos-13
             target: x86_64
@@ -143,7 +109,7 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.python-version }}
+          python-version: 3.x
       - name: Build wheels
         uses: PyO3/maturin-action@v1
         with:
