### 2025-02-28 92e866b
Reduce number of supported platforms for wheel building
