### 2025-02-28 3d3d8f6
Add QEMU setup as per maturin-action README
