### 2025-02-28 f974c5f
Remove QEMU setup
