### 2025-03-01 0988231
Add pytest to release workflow
