+++ tests/test_geometry.py
@@ -23,8 +23,6 @@ from pengwann.geometry import (
     identify_onsite_interactions,
     Site,
 )
-from pymatgen.analysis.structure_matcher import StructureMatcher
-from pymatgen.core import Structure
 
 
 def build_geometry(symbols: list[str]) -> Geometry:
@@ -164,20 +162,6 @@ def test_Geometry_wannier_assignments(geometry, data_regression) -> None:
 def test_Geometry_wannier_assignments_no_wann(geometry_no_x) -> None:
     with pytest.raises(ValueError):
         geometry_no_x.wannier_assignments
-
-
-def test_Geometry_as_structure(shared_datadir) -> None:
-    geometry = Geometry.from_xyz("wannier90", f"{shared_datadir}")
-    structure = geometry.as_structure()
-
-    with open(f"{shared_datadir}/geometry.json", "r") as stream:
-        serial = json.load(stream)
-
-    ref_structure = Structure.from_dict(serial)
-
-    sm = StructureMatcher()
-
-    assert sm.fit(structure, ref_structure)
 
 
 def test_identify_interatomic_interactions_elemental(
