+++ .github/workflows/release.yml
@@ -1,3 +1,8 @@
+# This file is autogenerated by maturin v1.8.2
+# To update, run
+#
+#    maturin generate-ci --pytest -o .github/workflows/test.yml github
+#
 name: Release
 
 on: push
@@ -37,6 +42,31 @@ jobs:
         with:
           name: wheels-linux-${{ matrix.platform.target }}
           path: dist
+      - name: pytest
+        if: ${{ startsWith(matrix.platform.target, 'x86_64') }}
+        shell: bash
+        run: |
+          set -e
+          python3 -m venv .venv
+          source .venv/bin/activate
+          pip install pengwann --find-links dist --force-reinstall
+          pip install pytest pytest-cov pytest-regressions pytest-random-order
+          pytest --ignore=tests/test_pmg_interface.py
+      - name: pytest
+        if: ${{ !startsWith(matrix.platform.target, 'x86') && matrix.platform.target != 'ppc64' }}
+        uses: uraimo/run-on-arch-action@v2
+        with:
+          arch: ${{ matrix.platform.target }}
+          distro: ubuntu22.04
+          githubToken: ${{ github.token }}
+          install: |
+            apt-get update
+            apt-get install -y --no-install-recommends python3 python3-pip
+            pip3 install -U pip pytest pytest-cov pytest-regressions pytest-random-order
+          run: |
+            set -e
+            pip3 install pengwann --find-links dist --force-reinstall
+            pytest --ignore=tests/test_pmg_interface.py
 
   musllinux:
     runs-on: ${{ matrix.platform.runner }}
@@ -68,6 +98,36 @@ jobs:
         with:
           name: wheels-musllinux-${{ matrix.platform.target }}
           path: dist
+      - name: pytest
+        if: ${{ startsWith(matrix.platform.target, 'x86_64') }}
+        uses: addnab/docker-run-action@v3
+        with:
+          image: alpine:latest
+          options: -v ${{ github.workspace }}:/io -w /io
+          run: |
+            set -e
+            apk add py3-pip py3-virtualenv
+            python3 -m virtualenv .venv
+            source .venv/bin/activate
+            pip install pengwann --no-index --find-links dist --force-reinstall
+            pip install pytest pytest-cov pytest-regressions pytest-random-order
+            pytest --ignore=tests/test_pmg_interface.py
+      - name: pytest
+        if: ${{ !startsWith(matrix.platform.target, 'x86') }}
+        uses: uraimo/run-on-arch-action@v2
+        with:
+          arch: ${{ matrix.platform.target }}
+          distro: alpine_latest
+          githubToken: ${{ github.token }}
+          install: |
+            apk add py3-virtualenv
+          run: |
+            set -e
+            python3 -m virtualenv .venv
+            source .venv/bin/activate
+            pip install pytest pytest-cov pytest-regressions pytest-random-order
+            pip install pengwann --find-links dist --force-reinstall
+            pytest --ignore=tests/test_pmg_interface.py
 
   windows:
     runs-on: ${{ matrix.platform.runner }}
@@ -95,6 +155,16 @@ jobs:
         with:
           name: wheels-windows-${{ matrix.platform.target }}
           path: dist
+      - name: pytest
+        if: ${{ !startsWith(matrix.platform.target, 'aarch64') }}
+        shell: bash
+        run: |
+          set -e
+          python3 -m venv .venv
+          source .venv/Scripts/activate
+          pip install pengwann --find-links dist --force-reinstall
+          pip install pytest pytest-cov pytest-regressions pytest-random-order
+          pytest --ignore=tests/test_pmg_interface.py
 
   macos:
     runs-on: ${{ matrix.platform.runner }}
@@ -121,6 +191,14 @@ jobs:
[…]
