### 2025-03-03 64cf30e
Revert "Test import separately from pytest"
This reverts commit 100a0e11759bbe2c2e6228f08160e47be1502d32.
