+++ rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pengwann"
-version = "0.4.0"
+version = "0.5.0"
 edition = "2024"
 
 [lib]
