### 2025-03-03 8c768c5
Remove redundant import in geometry tests
