+++ tests/test_geometry.py
@@ -13,7 +13,6 @@
 # You should have received a copy of the GNU General Public License along with pengWann.
 # If not, see <https://www.gnu.org/licenses/>.
 
-import json
 import pytest
 import numpy as np
 from pengwann.geometry import (
