### 2025-03-03 7fba8a3
Use --no-index to ensure wheel generation works without version bumping
