### 2025-03-03 a1963cc
Add tmate for debugging pytest in CI
