### 2025-03-03 d19d72d
Allow ssh to linux not windows
