+++ .github/workflows/release.yml
@@ -67,6 +67,9 @@ jobs:
             set -e
             pip3 install pengwann --find-links dist --force-reinstall
             pytest --ignore=tests/test_pmg_interface.py
+      - name: tmate
+        if: ${{ failure() }}
+        uses: mxschmitt/action-tmate@v3
 
   musllinux:
     runs-on: ${{ matrix.platform.runner }}
@@ -165,9 +168,6 @@ jobs:
           pip install pengwann --find-links dist --force-reinstall
           pip install pytest pytest-cov pytest-regressions pytest-random-order
           pytest --ignore=tests/test_pmg_interface.py
-      - name: tmate
-        if: ${{ failure() }}
-        uses: mxschmitt/action-tmate@v3
 
   macos:
     runs-on: ${{ matrix.platform.runner }}
