### 2025-03-03 76f7e6e
Finally fix (fingers crossed) pytest in maturin wheel CI
