### 2025-03-03 f00de11
Fix descriptors module tests
