### 2025-03-03 98161b4
Add action to update rust to latest stable in release CI
