### 2025-03-03 733781b
Remove redundant json import in descriptors module tests
