### 2025-03-04 de902aa
New CI model with wheels built and tested each commit
