+++ .github/workflows/ci.yml
@@ -1,10 +1,10 @@
-name: Continuous integration
+name: CI
 
-on: [push, pull_request]
+on: push
 
 jobs:
   check:
-    name: Lint, format and type check
+    name: Lint, format and validate types
     runs-on: ubuntu-latest
 
     steps:
@@ -14,7 +14,7 @@ jobs:
       - name: Install uv and Python
         uses: astral-sh/setup-uv@v5
         with:
-          python-version: "3.12"
+          python-version: 3.12
           enable-cache: true
 
       - name: Install Rust toolchain
@@ -22,33 +22,253 @@ jobs:
         with:
           components: clippy, rustfmt
 
-      - name: Sync environment
-        run: uv sync
-
       - name: Lint Rust
         working-directory: rust
         run: cargo clippy
 
-      - name: Enforce Rust formatting
+      - name: Check Rust formatting
         working-directory: rust
         run: cargo fmt --check
 
       - name: Lint Python
         run: uv run ruff check
 
-      - name: Enforce Python formatting
+      - name: Check Python formatting
         run: uv run ruff format --check
 
-      - name: Run static type checker
+      - name: Validate Python types
         run: uv run pyright
 
-  test:
-    name: Test and publish coverage
-    runs-on: ${{ matrix.os }}
+  linux-t1:
+    name: Build tier 1 Linux wheels
+    runs-on: ${{ matrix.platform.runner }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.10", "3.11", "3.12"]
+        platform:
+          - runner: ubuntu-latest
+            target: x86_64
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Install Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: 3.x
+
+      - name: Build wheels
+        uses: PyO3/maturin-action@v1
+        with:
+          target: ${{ matrix.platform.target }}
+          args: --release --out dist --find-interpreter
+          sccache: ${{ !startsWith(github.ref, 'refs/tags/') }}
+          manylinux: auto
+
+      - name: Upload wheels
+        uses: actions/upload-artifact@v4
+        with:
+          name: wheels-${{ runner.os }}-${{ matrix.platform.target }}
+          path: dist
+
+  linux-t2-glibc:
+    name: Build tier 2 Linux wheels (glibc)
+    runs-on: ${{ matrix.platform.runner }}
+    strategy:
+      matrix:
+        platform:
+          - runner: ubuntu-latest
+            target: x86
+          - runner: ubuntu-latest
+            target: aarch64
+          - runner: ubuntu-latest
+            target: armv7
+          - runner: ubuntu-latest
+            target: s390x
+          - runner: ubuntu-latest
+            target: ppc64le
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Install Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: 3.x
+
+      - name: Build wheels
+        uses: PyO3/maturin-action@v1
+        with:
+          target: ${{ matrix.platform.target }}
+          args: --release --out dist --find-interpreter
+          sccache: ${{ !startsWith(github.ref, 'refs/tags/') }}
+          manylinux: auto
+
+      - name: Upload wheels
+        uses: actions/upload-artifact@v4
+        with:
+          name: wheels-${{ runner.os }}-${{ matrix.platform.target }}
+          path: dist
+
+  linux-t2-musl:
+    name: Build tier 2 Linux wheels (musl)
+    runs-on: ${{ matrix.platform.runner }}
+    strategy:
+      matrix:
+        platform:
[…]
