### 2025-03-04 c9702a1
Fix python version matrix in test CI
