+++ .github/workflows/ci.yml
@@ -259,7 +259,7 @@ jobs:
     runs-on: ${{ matrix.platform.runner }}
     strategy:
       matrix:
-        python-version: [3.10, 3.11, 3.12]
+        python-version: ["3.10", "3.11", "3.12"]
         platform:
           - runner: ubuntu-latest
             target: x86_64
