+++ .github/workflows/test.yml
@@ -1,0 +1,58 @@
+name: Test wheels (tier 1)
+
+on: [pull_request, push, workflow_call]
+
+jobs:
+  check:
+    uses: ./.github/workflows/check.yml
+
+  build:
+    uses: ./.github/workflows/build-t1.yml
+
+  test:
+    needs: [check, build]
+    runs-on: ${{ matrix.platform.runner }}
+    strategy:
+      matrix:
+        python-version: ["3.10", "3.11", "3.12"]
+        platform:
+          - runner: ubuntu-latest
+            target: x86_64
+          - runner: macos-13
+            target: x86_64
+          - runner: macos-latest
+            target: aarch64
+          - runner: windows-latest
+            target: x64
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Install uv and Python
+        uses: astral-sh/setup-uv@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+          enable-cache: true
+
+      - name: Download wheels
+        uses: actions/download-artifact@v4
+        with:
+          name: wheels-${{ runner.os }}-${{ matrix.platform.target }}
+          path: dist
+
+      - name: Install wheel and test dependencies
+        run: |
+          uv venv
+          uv pip install pengwann --find-links dist
+          uv pip install pytest pytest-cov pytest-regressions pytest-random-order
+          uv pip install pymatgen
+
+      - name: Run test suite
+        run: uv run --no-sync pytest
+
+      - name: Publish test coverage
+        if: ${{ (github.event_name == 'push') && (runner.os == 'Linux') && (matrix.python-version == 3.12) }}
+        uses: paambaati/codeclimate-action@v9.0.0
+        env:
+          CC_TEST_REPORTER_ID: ${{ secrets.CODECLIMATE_REPO_TOKEN }}
