+++ .github/workflows/test.yml
@@ -1,6 +1,11 @@
 name: Test wheels
 
-on: [pull_request, push, workflow_call]
+on:
+  pull_request:
+  push:
+    branches:
+      - main
+  workflow_call:
 
 jobs:
   check:
