### 2025-03-04 59012aa
Attempt to exclude 3.13t wheels until next rust-numpy release
