### 2025-03-04 #7818 @JosePizarro3
Hi @PatrickJTaylor @rkurchin

Thanks for the opportunity, I will be very happy to review this package. The repo looks super-interesting 🙂

However, I am on parental leave until 26.03.2025. I will try to have some time to review this during the relaxed moments (like this one 😄), but cannot promise that I will finish very fast. So I leave it to both of you to decide whether you can wait 1 or 1.5 months for my review, or whether you prefer someone else to handle it.
