+++ .github/workflows/test.yml
@@ -1,24 +1,10 @@
-name: Test wheels (tier 1)
+name: test wheels (tier 1)
 
-on:
-  pull_request:
-  push:
-    branches:
-      - main
-  workflow_call:
+on: workflow_call
 
 jobs:
-  check:
-    name: Check
-    uses: ./.github/workflows/check.yml
-
-  build:
-    name: Build
-    uses: ./.github/workflows/build-t1.yml
-
   test:
-    name: Test
-    needs: [check, build]
+    name: 🖥 ${{ matrix.platform.runner }} - 🧬 ${{ matrix.platform.target }} / 🐍 ${{ matrix.python-version }}
     runs-on: ${{ matrix.platform.runner }}
     strategy:
       matrix:
