### 2025-03-05 bce447d
Update docs regarding rust extension
