+++ .readthedocs.yml
@@ -4,6 +4,7 @@ build:
   os: ubuntu-lts-latest
   tools:
     python: '3.12'
+    rust: '1.82'
 
 python:
   install:
