+++ .github/workflows/build-t1.yml
@@ -1,4 +1,4 @@
-name: build wheels (tier 1)
+name: build wheels (Tier 1)
 
 on: workflow_call
 
