+++ pyproject.toml
@@ -40,7 +40,8 @@ docs = [
   'numpydoc',
   'pydata-sphinx-theme',
   'sphinx',
-  'sphinxcontrib-bibtex'
+  'sphinxcontrib-bibtex',
+  'tomli',
 ]
 
 [project.urls]
@@ -85,3 +86,4 @@ dev = [
     "scipy-stubs>=1.15.1.0",
     "types-tqdm>=4.67.0.20241221",
 ]
+docs = []
