### 2025-03-05 #7818 @PatrickJTaylor
👋 @rkurchin @JosePizarro3

Firstly let me say thank you @rkurchin for taking on this submission so promptly - I know the JOSS backlog is quite hefty at the minute so I really appreciate it 😃

@JosePizarro3 thank you for your offer to review the package! I appreciate the transparency re: the rough timeline, personally I am happy to wait ~ 1/1.5 months if this is also okay with you @rkurchin?
