### 2025-03-05 593c890
Move pmg test back into geometry tests
