### 2025-03-05 861ac0c
Rust edition 2024 => 2021 for rtd compatibility
