+++ rust/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "pengwann"
 version = "0.4.0"
-edition = "2024"
+edition = "2021"
 
 [lib]
 name = "pengwann"
