### 2025-03-05 e6ea766
Fix rust formatting for 2021 edition
