### 2025-03-05 6d29a28
Allow test workflow to inherit codeclimate secret
