### 2025-03-06 987743a
Add basic docstrings to PyO3 module
