### 2025-03-06 001c1d0
Updates to installation and contributing guides for Rust extension
