### 2025-03-06 1b01e44
Use maturin rather than uv to publish to PyPI
