### 2025-03-06 3c58e0a
Add GPL comments to rust source
