+++ rust/src/lib.rs
@@ -3,6 +3,21 @@
 //! Provides a few Rust implementations for functions that would be prohibitively slow
 //! in Python, mainly due to loops that cannot be unrolled by the interpreter.
 
+// Copyright (C) 2024-2025 Patrick J. Taylor
+
+// This file is part of pengWann.
+//
+// pengWann is free software: you can redistribute it and/or modify it under the terms
+// of the GNU General Public License as published by the Free Software Foundation, either
+// version 3 of the License, or (at your option) any later version.
+//
+// pengWann is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+// without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+// PURPOSE. See the GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License along with pengWann.
+// If not, see <https://www.gnu.org/licenses/>.
+
 use std::iter::repeat_n;
 
 use itertools::Itertools;
