+++ rust/src/python_module.rs
@@ -1,5 +1,20 @@
 //! A PyO3 module that allows Python to call the Rust extension.
 
+// Copyright (C) 2024-2025 Patrick J. Taylor
+
+// This file is part of pengWann.
+//
+// pengWann is free software: you can redistribute it and/or modify it under the terms
+// of the GNU General Public License as published by the Free Software Foundation, either
+// version 3 of the License, or (at your option) any later version.
+//
+// pengWann is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+// without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+// PURPOSE. See the GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License along with pengWann.
+// If not, see <https://www.gnu.org/licenses/>.
+
 use numpy::{IntoPyArray, PyArray2, PyArray3, PyReadonlyArray2};
 use pyo3::prelude::*;
 
