### 2025-03-06 d2acea9
Remove logos from license and PyPI badges
