### 2025-03-06 5d131bb
Remove Ruff badge (for sake of visual clutter)
