### 2025-03-07 0be7a71
Sync badges on README and docs index page
