### 2025-03-07 6be8bcc
No longer need to omit now-deleted version.py from coverage
