### 2025-03-07 03669e3
Bump version in CITATION.cff
