### 2025-03-10 65dd80c
Re-enable testing for 3.13
