### 2025-03-11 #7818 @rkurchin
👋 @xwang862 and/or @LIVazquezS, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
