+++ rust/Cargo.toml
@@ -11,5 +11,5 @@ crate-type = ["cdylib", "rlib"]
 itertools = "0.14.0"
 ndarray = "0.16.1"
 ndarray-stats = "0.6.0"
-numpy = "0.23.0"
-pyo3 = { version = "0.23.5", features = ["extension-module"] }
+numpy = "0.24.0"
+pyo3 = { version = "0.24.0", features = ["extension-module"] }
