### 2025-03-18 f460f4e
Switch from pyright -> basedpyright
