### 2025-03-18 c2e3adf
First pass improving typing with basedpyright
