+++ src/pengwann/io/wannier90.py
@@ -25,6 +25,7 @@ parsing all the data required to construct an instance of the
 from __future__ import annotations
 
 import os
+
 import numpy as np
 from numpy.typing import NDArray
 from scipy.constants import physical_constants
@@ -108,24 +109,19 @@ def read_eigenvalues(
     eigenvalues : ndarray of float
         The Kohn-Sham eigenvalues.
     """
-    eigenvalues_list = []
-
     with open(path, "r") as stream:
         lines = stream.readlines()
 
-    block_indices = [idx * num_bands for idx in range(num_kpoints)]
+    eigenvalues = np.zeros((num_bands, num_kpoints))
 
-    for column_idx in range(num_bands):
-        row = []
+    n_lines = range(num_bands)
+    k_lines = [idx * num_bands for idx in range(num_kpoints)]
 
-        for block_idx in block_indices:
-            eigenvalue = float(lines[column_idx + block_idx].split()[-1])
+    for n, n_line in enumerate(n_lines):
+        for k, k_line in enumerate(k_lines):
+            eigenvalue = float(lines[n_line + k_line].split()[-1])
 
-            row.append(eigenvalue)
-
-        eigenvalues_list.append(row)
-
-    eigenvalues = np.array(eigenvalues_list)
+            eigenvalues[n, k] = eigenvalue
 
     return eigenvalues
 
@@ -146,45 +142,33 @@ def read_u(path: str) -> tuple[NDArray[np.complex128], NDArray[np.float64]]:
     kpoints : ndarray of float
         The k-point mesh used in the prior Wannier90 calculation.
     """
-    u_list, kpoints_list = [], []
-
     with open(path, "r") as stream:
         lines = stream.readlines()
 
     num_kpoints, num_wann, num_bands = [int(string) for string in lines[1].split()]
 
-    block_indices = (idx * (num_wann * num_bands + 2) + 4 for idx in range(num_kpoints))
-    row_indices = [idx for idx in range(num_bands)]
-    column_indices = [idx * num_bands for idx in range(num_wann)]
+    u = np.zeros((num_kpoints, num_bands, num_wann), dtype=np.complex128)
+    kpoints = np.zeros((num_kpoints, 3))
 
-    for block_idx in block_indices:
-        u_k = []
+    k_lines = (idx * (num_wann * num_bands + 2) + 4 for idx in range(num_kpoints))
+    n_lines = [idx for idx in range(num_bands)]
+    w_lines = [idx * num_bands for idx in range(num_wann)]
 
-        kpoint = [float(string) for string in lines[block_idx - 1].split()]
-        kpoints_list.append(kpoint)
+    for k, k_line in enumerate(k_lines):
+        kpoints[k] = [float(string) for string in lines[k_line - 1].split()]
 
-        for row_idx in row_indices:
-            row = []
-
-            for column_idx in column_indices:
-                element_idx = block_idx + row_idx + column_idx
+        for n, n_line in enumerate(n_lines):
+            for w, w_line in enumerate(w_lines):
                 real, imaginary = [
-                    float(string) for string in lines[element_idx].split()
+                    float(string) for string in lines[k_line + n_line + w_line].split()
                 ]
 
-                row.append(complex(real, imaginary))
-
-            u_k.append(row)
-
-        u_list.append(u_k)
-
-    u = np.array(u_list)
-    kpoints = np.array(kpoints_list)
+                u[k, n, w] = complex(real, imaginary)
 
     return u, kpoints
 
 
-def read_hamiltonian(path: str) -> dict[tuple[int, ...], NDArray[np.complex128]]:
[…]
