### 2025-03-18 a764a2b
Silence basedpyright wrt private usage re: rust
