### 2025-03-18 d2aaf9a
Do not use stubPath to allow io module
