+++ tests/test_utils.py
@@ -15,7 +15,7 @@
 
 import numpy as np
 
-from pengwann.io.wannier90 import read_u
+from pengwann.io import read_u
 from pengwann.utils import (
     get_spilling_factor,
     integrate_descriptor,
