### 2025-03-18 156152d
Update CONTRIBUTING re: basedpyright
