### 2025-03-18 bf11972
Narrow type of H keys with additional assert
