### 2025-03-19 517bde7
Point basedpyright to 3.10 re: typing support
