### 2025-03-19 4949d6b
Add Self from typing_extensions
