+++ src/pengwann/geometry.py
@@ -29,7 +29,7 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 from numpy.typing import ArrayLike, NDArray
-from typing_extensions import override
+from typing_extensions import Self, override
 
 from pengwann._geometry import (
     _build_distance_and_image_matrices,  # pyright:ignore[reportPrivateUsage]
@@ -238,7 +238,7 @@ class Geometry:
     @classmethod
     def from_xyz(
         cls, seedname: str, path: str = ".", cell: ArrayLike | None = None
-    ) -> Geometry:
+    ) -> Self:
         """
         Initialise a Geometry object from a seedname_centres.xyz file.
 
