### 2025-03-19 3c9dda3
Fix docstring type for image_matrix i32
