### 2025-03-19 5cc04ae
Make assign_h_ij private and add error handling
