### 2025-03-19 132e4f5
Remove now redundant test
