### 2025-03-19 73fb6ff
Add error handling if unable to determine n_cpus
