### 2025-04-07 83463d0
Explicitly point out local examples
