### 2025-04-07 dbd0b42
Use absolute link to example for PyPI compatibility
