### 2025-04-07 5372e9f
maturin-action => 1.48.0, attempt to fix t2 builds
