### 2025-04-07 94ae14a
Explicitly compare LOBSTER and pengwann spilling in paper
