### 2025-04-07 a4fcac1
Improve phrasing of paper footnote
