### 2025-04-07 debcc6b
Swap introduction of eigenvalues and eigenvectors
