### 2025-04-07 736e81f
Move ht_scdm and pwdf references to highlight meaning of problematic sentence
