+++ paper/paper.md
@@ -68,7 +68,7 @@ More generally, even in systems where pre-defined atomic or pseudo-atomic orbita
 The LOBSTER code [@lobster_2016;@lobster_2020] implements much of the same functionality as `pengwann` using basis sets of pre-defined atomic and pseudo-atomic orbitals. LOBSTER offers additional features not directly supported by `pengwann`, such as generating fatband plots and obtaining localised molecular orbitals via transformation of the projected atomic orbital basis [@lobster_fragment]. We anticipate that most potential users of `pengwann` will already be familiar with LOBSTER, warranting a brief discussion of the relative advantages and disadvantages of each code.
 
 `pengwann` achieves lower spilling factors than LOBSTER[^1] and readily handles systems when non-atom-centered basis functions are preferable.
-Moreover, since Wannier functions serve many purposes in addition to the calculation of bonding descriptors, they may already have been calculated for systems of interest, making the use of `pengwann` particularly efficient in these cases. Conversely, LOBSTER's main advantage lies in its pre-defined atomic and pseudo-atomic basis sets, which can be applied to any system with minimal user input, while obtaining high-quality Wannier functions can be complex due to their strong non-uniqueness [@wannier_review]. However, recent advances have significantly simplified the computation of well-localised Wannier functions with appropriate symmetry for chemical bonding analysis, lowering the barrier to their use [@ht_scdm;@pwdf].
+Moreover, since Wannier functions serve many purposes in addition to the calculation of bonding descriptors, they may already have been calculated for systems of interest, making the use of `pengwann` particularly efficient in these cases. Conversely, LOBSTER's main advantage lies in its pre-defined atomic and pseudo-atomic basis sets, which can be applied to any system with minimal user input, while obtaining high-quality Wannier functions can be complex due to their strong non-uniqueness [@wannier_review]. However, recent advances [@ht_scdm;@pwdf] have significantly simplified the computation of well-localised Wannier functions with appropriate symmetry for chemical bonding analysis, lowering the barrier to their use.
 
 The WOBSTER code [@wobster] implements a subset of the features found in `pengwann`, allowing users to compute the Wannier-projected density of states and the Wannier orbital Hamilton population. `pengwann` provides a broader set of features than WOBSTER and is more performant: `pengwann` leverages `numpy` [@numpy] and the built-in `multiprocessing` library to vectorise and parallelise most operations, whereas WOBSTER's loop-based approach is considerably slower.
 
