### 2025-04-07 f2d79a2
Add more specific details to comparison with WOBSTER
