### 2025-04-07 #7890 @PatrickJTaylor
@LIVazquezS - thank you kindly for your review of the code and the paper, I am glad to hear that you enjoyed looking through the docs and trying out the package!

To briefly respond to the comments that you raised above:
- The example notebooks in the docs can already be executed locally, but as you pointed out, this was not previously made obvious. I have now explicitly highlighted the possibility of playing around with the example notebooks and executing them locally in the README (PatrickJTaylor/pengWann#25)
- Re: evidence for `pengwann` achieving lower spilling factors than LOBSTER, I have added a small footnote in the paper to explicitly highlight some examples of this (PatrickJTaylor/pengWann#26). I should also reiterate that as mentioned in the paper, the suggestion that `pengwann` should achieve lower spilling factors is not based solely on empirical observations, it is also (at least for systems with energetically isolated bands) a mathematical certainty that the spilling factor should be strictly 0 (within machine precision of course), which I would say is itself quite strong "evidence" of our claim.
- More generally, I have made several minor edits to the manuscript in response to your remaining comments (PatrickJTaylor/pengWann#27) - thank you for highlighting areas that were not as clear as they could be.
- You mentioned that it might be best if we explicitly introduce the formalism of Wannier functions in the paper, with this already being featured in parts of the documentation. Whilst I agree that this would be ideal, I am not sure that we could include the same level of detail in the paper as can be found in the docs without substantially increasing the length of the paper, which is already above the approximate JOSS limit of 1000 words (one of the reasons that I chose to include a distinct methodology section in the docs was precisely because I didn't feel we had space for it in the JOSS paper). I suppose I could include literally just the basic definition of a Wannier function without adding too many extra words, but I am not sure how much this adds for readers who are already unfamiliar with the subject matter - all the same I am happy to add this if you think it would be beneficial.
- In terms of a "graphical description" of the code's functions, do you mean that we should add a few example output figures? This would be an easy change, I just avoided any figures initially for the sake of brevity (and because there are plenty of examples in the docs).
- Finally, you mention that it would be good to improve the installation instructions with respect to conda users, do you mean that we should maintain a separate conda package (so that you can do `conda install pengwann`) or just explicitly mention that `conda` users can install `pengwann` in much the same way as in any other Python virtual environment? In principle I would be happy to develop and maintain a `conda` recipe for `pengwann`, but as I have not done this before I suspect it would take me some time to get it working for all supported architectures etc.

Thanks again for your review, I really appreciate it 😃 - I would suggest that if you have a lot more to add with regards to all of the above, feel free to open some issues in the pengWann repo and we can discuss further over there.
