### 2025-04-08 #7890 @LIVazquezS
Hi @PatrickJTaylor,

Thank you so much for the prompt response! I appreciate your effort and the great quality of your software.  Most of the points, I raised, were solved. Just a couple of general comments.

> You mentioned that it might be best if we explicitly introduce the formalism of Wannier functions in the paper, with this already being featured in parts of the documentation. Whilst I agree that this would be ideal, I am not sure that we could include the same level of detail in the paper as can be found in the docs without substantially increasing the length of the paper, which is already above the approximate JOSS limit of 1000 words (one of the reasons that I chose to include a distinct methodology section in the docs was precisely because I didn't feel we had space for it in the JOSS paper). I suppose I could include literally just the basic definition of a Wannier function without adding too many extra words, but I am not sure how much this adds for readers who are already unfamiliar with the subject matter - all the same I am happy to add this if you think it would be beneficial.

I am happy with your explanation. However, and for sake of completeness, I suggest adding a line pointing the reader to the documentation with the full theoretical explanation.

> In terms of a "graphical description" of the code's functions, do you mean that we should add a few example output figures? This would be an easy change, I just avoided any figures initially for the sake of brevity (and because there are plenty of examples in the docs).

Indeed! That is precisely what I meant. Nowadays, most people would quickly screen a paper and look at the figures if they show what they are looking for, so I think a brief illustration of the main capabilities of the code would be a great addition. Although, if you would like to keep the paper as it is, pointing the reader to the documentation should be enough.

> Finally, you mention that it would be good to improve the installation instructions with respect to conda users, do you mean that we should maintain a separate conda package (so that you can do conda install pengwann) or just explicitly mention that conda users can install pengwann in much the same way as in any other Python virtual environment? In principle I would be happy to develop and maintain a conda recipe for pengwann, but as I have not done this before I suspect it would take me some time to get it working for all supported architectures etc.

I think both options are valid, and I leave it to your consideration. I understand that having a package in ´conda´ could be a pain, and a competent developer could easily install the package in `conda`from the existent instructions.

After those simple changes, I would be happy to suggest acceptance of your paper. Thanks again for your kind answers and your nice work. I would look forward to further developments of `pengwann`.
