### 2025-04-08 15a712b
Add link to docs from JOSS paper
