### 2025-04-09 d7a0e63
Manually add LICENSE to maturin sdist
