+++ rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pengwann"
-version = "0.5.0"
+version = "0.5.1"
 edition = "2021"
 
 [lib]
