### 2025-04-10 2bb7b60
Add conda badge to README and docs
