### 2025-04-17 #7890 @JosePizarro3
Hi there, I finished my review!

Overall, the code is very nice. You followed a very nice structure, and you have created a very clear release ecosystem, so congrats! It is not so easy to find such projects when coming to scientific software out there :-)

**I'd like to recommend `pengWann` as a publication in JOSS**

Everything else worked, and the documentation, examples, and paper are easy to read. I agree with all comments from @LIVazquezS as well. Nevertheless, I have some minor comments or suggestions (feel free to accept them or not, as you wish), even if they read as orders or sound too hard, please, bear in mind these are all recommendations, and I just want to get to the point without using long and tiring expressions.

## Code

- You can add `uv` in the installation steps for a quicker and cleaner installation.
- You could think of taking out `typing-extensions` as a dependency of the pyproject.toml. I understand that typing is useful, but I think overall, you reached the point when typing becomes unreadable. This happened to me while reading a bit your src code (especially when using `Self`, `NDArray`, `tuple[tuple[tuple]]]` and these kind of nested typing expressions). You could relax it; at the end, typing should not make your functions complicated to read, and you can always use the docstrings to be more expressive on what the i/o of the functions mean.

### Examples

- What about using `{seedname}.wout` to parse the info from Wannier90? Does it contain the XYZ and cell info right?
- I also suggest never using the input files when coming to parse. These are prone to be modified, and it can cause a lot of troubles.
- You could also store the parsed data into a dataclass (or a pedantic BaseModel), whose fields are the data. Example, when you extract in example 1 the `kpoints`, `eigenvalues`, `u_matrices`, `hamiltonian`, these could be fields of a dataclass called `Wannier90Data`
- The reading of the `.eig` when the ab initio code data is not available, is done from the Wannier90 data. I think the phrasing is incorrect in example 1
- Lastly, regarding parsing data, I think is important to have a way to connect Wannier90 with whatever (VASP, QE...) ab initio code that was initially performed. Do you control that anyway? I could not find it in the src code, but I might have overlooked it. This might be a suggestion for a future major release.

## Documentation

- You could think of structuring the docu a bit differently. For a future release, maybe you want to read about [diataxis](https://diataxis.fr/)
- I'd also avoid writing methodology text in the documentation. I think I've never read this section **ever**, as I always thought there are way better sources of information to get this. I would only agree to show this in cases where the actual implementation of definitions is shown, and not even in 100% of the cases. The main question here is how long did it take you to write it, and whether it is worth for the time spent (i.e., how many users will actually use the readthedocs methodology section).

## Paper

(very cherry picked comments)
- I would simply delete "quantum chemistry" from the abstract, and use "First-principles calculations". Especially because Chemistry is not there to address periodic systems, but Condensed Matter Physics is 😄
- I will express different "some Hamiltonian" in the abstract. Maybe use more specific wording like "many-body Hamiltonian". I mean, it is a very specific Hamiltonian the one that is solved in any quantum chemistry / condensed matter physics method, i.e., the atomistic one.

---

Thanks again for the hard work, @PatrickJTaylor, the repo is very interesting, very welcoming to read and to use, and I am definitely motivated to give it a try :-)
