### 2025-04-17 5042243
Read cell vectors from wout, not win
