### 2025-04-17 35d8d27
Update geometry module for new read_cell impl
