+++ tests/test_io.py
@@ -13,8 +13,6 @@
 # You should have received a copy of the GNU General Public License along with pengWann.
 # If not, see <https://www.gnu.org/licenses/>.
 
-import pytest
-
 from pengwann.io import (
     read,
     read_cell,
