### 2025-04-17 51b59d2
Remove quantum chemistry phrasing from paper
