### 2025-04-17 97939ad
Some H -> many-body H
