### 2025-04-17 7331e10
Explicitly install 3.10 on macos runners
