+++ .github/workflows/build-t1.yml
@@ -45,6 +45,11 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4
 
+      - name: Install Python
+        uses: actions/setup-python@8d9ed9ac5c53483de85588cdf95a591a75ab9f55 # v5
+        with:
+          python-version: "3.10"
+
       - name: Install Rust toolchain
         uses: dtolnay/rust-toolchain@stable
 
