### 2025-04-17 05f79be
Explicitly install python interpreters for all platforms
