### 2025-04-17 0042ba4
Remove redundant setup-python for manylinux builds
