### 2025-04-17 #7890 @PatrickJTaylor
@JosePizarro3 - thank you very much for your review, I really appreciate you taking the time to look over everything!

To briefly respond to some of your suggestions:

> You can add uv in the installation steps for a quicker and cleaner installation.

I have used `uv` primarily as a tool for development, I have not thought of it so much as something for the end user. There are already instructions for setting up a suitable development environment using `uv` in the contributing guide, which is also linked from the full installation guide in the docs, so I think I will leave it at that for now (though I appreciate you bringing this point to my attention all the same!)

> You could think of taking out typing-extensions as a dependency of the pyproject.toml. I understand that typing is useful, but I think overall, you reached the point when typing becomes unreadable. This happened to me while reading a bit your src code (especially when using Self, NDArray, tuple[tuple[tuple]]] and these kind of nested typing expressions). You could relax it; at the end, typing should not make your functions complicated to read, and you can always use the docstrings to be more expressive on what the i/o of the functions mean.

This is definitely a fair criticism, there is no doubt that the more complicated type hints make the source code less readable. I am hesitant to make the type hinting any less specific though, as I find there are many benefits to strong and static type systems. I could improve the situation somewhat without making the type hints any less specific by making judicious use of type aliases. This should make the source a little more readable to someone who is just giving the code a quick once-over to try and understand a specific implementation detail.

Despite the fact that I largely agree with your suggestions on this, I will not be addressing this issue as part of the review process, as I think I need more time to find a better solution.

> What about using {seedname}.wout to parse the info from Wannier90? Does it contain the XYZ and cell info right?

> I also suggest never using the input files when coming to parse. These are prone to be modified, and it can cause a lot of troubles.

I totally agree with this, I think I must have somehow forgotten that the cell vectors are printed in seedname.wout during developement! I have merged a PR to address this: PatrickJTaylor/pengWann#37.

> You could also store the parsed data into a dataclass (or a pydantic BaseModel), whose fields are the data. Example, when you extract in example 1 the kpoints, eigenvalues, u_matrices, hamiltonian, these could be fields of a dataclass called Wannier90Data

I agree that the current `io` implementation is not the most intuitive. I think rather than implementing a `Wannier90Data` dataclass, I would be more tempted to add a `from_files` classmethod to the `DescriptorCalculator`, but I need more time to ponder this one so I will not be implementing any changes as part of this review.

> You could think of structuring the docu a bit differently. For a future release, maybe you want to read about [diataxis](https://diataxis.fr/)

I am definitely keen to further improve the documentation, thank you for pointing out diataxis! I think I will have a proper look over all of the possible documentation options at some point, for the sake of future releases, but for now I will stick with what I have.

> I'd also avoid writing methodology text in the documentation. I think I've never read this section ever, as I always thought there are way better sources of information to get this. I would only agree to show this in cases where the actual implementation of definitions is shown, and not even in 100% of the cases. The main question here is how long did it take you to write it, and whether it is worth for the time spent (i.e., how many users will actually use the readthedocs methodology section).

Hahaha that is a very good point, probably many potential users will ignore the methodology section (or will not even find it in the first place). I mainly wrote it because a) I enjoy explaining aspects of theory and b) it is a more concise read (than the relevant papers) for those who just want to understand enough to know what they are actually doing when they use the software. Thankfully it also didn't take me very long to write, as there was already a very similar section in my PhD thesis!

> I would simply delete "quantum chemistry" from the abstract, and use "First-principles calculations". Especially because Chemistry is not there to address periodic systems, but Condensed Matter Physics is 😄

> I will express different "some Hamiltonian" in the abstract. Maybe use more specific wording like "many-body Hamiltonian". I mean, it is a very specific Hamiltonian the one that is solved in any quantum chemistry / condensed matter physics method, i.e., the atomistic one.

I think these are both great points and I have addressed both of them in PatrickJTaylor/pengWann#38

Thank you again for the time and effort you spared to write your review, I really appreciate all of the comments and I hope `pengwann` proves useful to you should you give it a proper try in the future!
