### 2025-04-17 #7890 @rkurchin
Editorial comments...

### paper
- line 27: suggest adding a colon at the end of this line
- footnote 1: to my eye, scientific notation with a percent looks a bit odd and unintuitive. It's up to you, but I would change all the spilling factors to just be raw numbers and dispense with the % signs (unless always reporting it as percent is a strong convention in this community, to which I am certainly closely adjacent in my own work but not directly a part of)

### references
- WOBSTER has a citation, see [instructions from README](https://github.com/Chengcheng-Xiao/WOBSTER?tab=readme-ov-file#how-to-cite)
