### 2025-04-17 a09b6fd
Use raw values for spilling factors
