### 2025-04-17 d6c2889
Add explicit paper reference for WOBSTER
