+++ paper/paper.md
@@ -72,7 +72,7 @@ The LOBSTER code [@lobster_2016;@lobster_2020] implements much of the same funct
 `pengwann` achieves lower spilling factors than LOBSTER[^1] and readily handles systems when non-atom-centered basis functions are preferable.
 Moreover, since Wannier functions serve many purposes in addition to the calculation of bonding descriptors, they may already have been calculated for systems of interest, making the use of `pengwann` particularly efficient in these cases. Conversely, LOBSTER's main advantage lies in its pre-defined atomic and pseudo-atomic basis sets, which can be applied to any system with minimal user input, while obtaining high-quality Wannier functions can be complex due to their strong non-uniqueness [@wannier_review]. However, recent advances such as those of Refs. [@ht_scdm;@pwdf] have significantly simplified the computation of well-localised Wannier functions with appropriate symmetry for chemical bonding analysis, lowering the barrier to their use.
 
-The WOBSTER code [@wobster] implements a subset of the features found in `pengwann`, allowing users to compute the Wannier-projected density of states and the Wannier orbital Hamilton population. `pengwann` provides a broader set of features than WOBSTER and is more performant: `pengwann` leverages `numpy` [@numpy] and the built-in `multiprocessing` library to vectorise and parallelise most operations, whereas WOBSTER's loop-based approach is considerably slower. To give a specific example: the Wannier orbital Hamilton population for a given pair of Wannier functions is computed in WOBSTER using a triple nested loop over energies, k-points and bands. The same calculation in `pengwann` is fully vectorised across these three dimensions and furthermore many such computations can be carried out in parallel for different pairs of Wannier functions.
+The WOBSTER code [@wobster;@wannier_population_analysis] implements a subset of the features found in `pengwann`, allowing users to compute the Wannier-projected density of states and the Wannier orbital Hamilton population. `pengwann` provides a broader set of features than WOBSTER and is more performant: `pengwann` leverages `numpy` [@numpy] and the built-in `multiprocessing` library to vectorise and parallelise most operations, whereas WOBSTER's loop-based approach is considerably slower. To give a specific example: the Wannier orbital Hamilton population for a given pair of Wannier functions is computed in WOBSTER using a triple nested loop over energies, k-points and bands. The same calculation in `pengwann` is fully vectorised across these three dimensions and furthermore many such computations can be carried out in parallel for different pairs of Wannier functions.
 
 # Acknowledgements
 
