+++ CITATION.cff
@@ -13,5 +13,5 @@ license: GPL-3.0
 message: "If you use this software, please cite it as below."
 repository-code: "https://github.com/PatrickJTaylor/pengWann"
 title: "pengWann"
-version: 0.5.1
+version: 0.5.2
 url: "https://pengwann.readthedocs.io"
