+++ .github/workflows/release.yml
@@ -35,7 +35,7 @@ jobs:
 
     steps:
       - name: Download sdist and wheels
-        uses: actions/download-artifact@95815c38cf2ff2164869cbab79da8d1f422bc89e # v4
+        uses: actions/download-artifact@d3f86a106a0bac45b974a628896c90dbdf5c8093 # v4
 
       - name: Publish to PyPI
         uses: PyO3/maturin-action@aef21716ff3dcae8a1c301d23ec3e4446972a6e3 # v1
