### 2025-04-25 d7b1dbf
Update astral-sh/setup-uv action to v6
