### 2025-04-28 #7890 @Kevin-Mattheus-Moerman
👉 version tag listed here should be set to `v0.5.2` to include the `v`.
